feat(auth): split login pages into Google login and test-code login; add session endpoint
